Open buffers in existing ERC window, if possible
Sometimes (very often in Emacs 26.1), `switch-to-buffer-other-window`
will split an existing ERC window. This package is supposed to provide
a hexchat-like experience for ERC users, and hexchat only has the
equivalent of one chat window.

Additional windows can be created manually, but we'll change the
`display-buffer` action to look for/re-use an existing ERC buffer on
the current frame to switch when selecting clicking on a channel name.