Prepare for inclusion in ERC (part of GNU Emacs)
* erc-status-sidebar.el: Indicate the Free Software Foundation as the
copyright holder, and say that this file is now part of GNU Emacs.
Tweak the Commentary to more closely follow the style used throughout
the rest of GNU Emacs.  Shorten some of the longer lines, and tweak
comments to adhere to the preferred style.  See the CONTRIBUTE file in
GNU Emacs's source repository for more details.